% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_cols.R
\name{add_missing_cols}
\alias{add_missing_cols}
\title{Append columns to matrix x corresponding to columns of matrix y}
\usage{
add_missing_cols(x, colnames_y)
}
\arguments{
\item{x}{a matrix (missing rows corresponding to matrix y)}

\item{colnames_y}{vector of colnames of matrix y}
}
\value{
matrix x with missing cols appended (entries are NA)
}
\description{
Append columns to matrix x corresponding to columns of matrix y
}
\examples{

}
