% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_rows.R
\name{add_missing_rows}
\alias{add_missing_rows}
\title{Append rows to matrix x corresponding to rows of matrix y}
\usage{
add_missing_rows(x, rownames_y)
}
\arguments{
\item{x}{a matrix (missing rows corresponding to matrix y)}

\item{rownames_y}{vector of rownames of matrix y}
}
\value{
matrix x with missing rows appended (entries are NA)
}
\description{
Append rows to matrix x corresponding to rows of matrix y
}
\examples{

}
