% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_matrix_margins.R
\name{align_matrix_margins}
\alias{align_matrix_margins}
\title{Align matrix margins}
\usage{
align_matrix_margins(mat_list, na_rows = TRUE, na_cols = FALSE,
  use_dims = NULL)
}
\arguments{
\item{mat_list}{list of matrices (columns are genes, rows are samples)}

\item{na_rows}{if TRUE, add NAs for missing rows, otherwise they are removed}

\item{na_cols}{if TRUE, add NAs for missing columns, otherwise they are removed}

\item{use_dims}{either an integer or character vector of length 1}
}
\value{
new list of matrices
}
\description{
Align matrix margins
}
\examples{

}
