% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{ari_score}
\alias{ari_score}
\title{Ari selectivity score for PRISM profile
 @export ari_score
 @param x viability, must be in 0 ~ 100 format, i.e. capped from 0 to 100}
\usage{
ari_score(x)
}
\description{
Ari selectivity score for PRISM profile
 @export ari_score
 @param x viability, must be in 0 ~ 100 format, i.e. capped from 0 to 100
}
