% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_render.r
\name{cds_render}
\alias{cds_render}
\title{Render a RMarkdown file and upload it to https://dev.cds.team/cds_projects}
\usage{
cds_render(path_to_rmd, folder = NULL, path_to_html = NULL, ...)
}
\arguments{
\item{path_to_rmd}{The path to the RMarkdown file you would want to render and upload to cds_projects}

\item{folder}{Folder to store your file into. Can't be an already Git taken folder name}

\item{path_to_html}{The path to the HTML File you already generated on your owm. We will skip the rendering of the RMD, so please make sure both match}

\item{...}{All parameteres you would want to pass to the rmarkdown render function}
}
\value{
The full path of the HTML generated
}
\description{
Render a RMarkdown file and upload it to https://dev.cds.team/cds_projects
}
\examples{
cds_render("local_rmd.Rmd")
}
