% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows_of_matrix.R
\name{collapse_rows_of_matrix}
\alias{collapse_rows_of_matrix}
\title{Collapse rows of matrix according to grouping variable}
\usage{
collapse_rows_of_matrix(mat, group_df, collapse_fun = Matrix::colMeans,
  group_var = Group, sample_var = Sample, do_parallel = F, ...)
}
\arguments{
\item{mat}{input matrix}

\item{group_df}{mapping from sample_var (row names of matrix) to specified group_var}

\item{collapse_fun}{function to collapse samples into groups}

\item{group_var}{grouping variable}

\item{sample_var}{sample variable; values must match row names of matrix}

\item{do_parallel}{T or F; parallelize collapse operation}

\item{...}{additional arguments to `collapse_fun`}
}
\value{
matrix with collapsed values
}
\description{
Collapse rows of matrix according to grouping variable
}
\examples{
nrow <- 100
ncol <- 20
x <- rnorm(nrow*ncol) \%>\% matrix(nrow=nrow, ncol=ncol) \%>\%
  magrittr::set_rownames(str_c("Guide_", 1:nrow)) \%>\%
  magrittr::set_colnames(str_c("CL", 1:ncol))
df <- data_frame(GuideID=rownames(x),
                 GeneID=sample(str_c("Gene_", LETTERS), nrow, replace=T))
res <- cdsr::collapse_rows_of_matrix(x, df, group_var=GeneID, sample_var=GuideID,
                               collapse_fun=Matrix::colMeans, na.rm=T)
}
