% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{compute_FCPC}
\alias{compute_FCPC}
\title{general function to compute FCPC
removing the top2 doses}
\usage{
compute_FCPC(df_mfi)
}
\arguments{
\item{df_mfi}{data frame containing columns of mfi, det_plate, pool_id, cell_line_id}
}
\description{
general function to compute FCPC
removing the top2 doses
}
