% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_test.R
\name{corr_test}
\alias{corr_test}
\title{Calculate correlations between matrices, with significance measures}
\usage{
corr_test(x, y, adjust)
}
\arguments{
\item{x}{matrix of values}

\item{y}{second matrix of values}

\item{adjust}{p-value correction method (fed into p.adjust).}
}
\value{
: list of matrices with elements r: correlations, n: number observations, t: t-stats, p: p-values, se: standard errors, q: q-values
}
\description{
Calculate correlations between matrices, with significance measures
}
