% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{dbPRISM_get_assaydata}
\alias{dbPRISM_get_assaydata}
\title{query dbPRISM for compound assay data}
\usage{
dbPRISM_get_assaydata(Broad_ID, project_ID)
}
\description{
query dbPRISM for compound assay data
}
