% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sparse_matrix.R
\name{download_sparse_matrix}
\alias{download_sparse_matrix}
\title{download raw sparse matrix data files from taiga to R}
\usage{
download_sparse_matrix(data_name, version_number, file_name)
}
\arguments{
\item{data_name}{the name of the taiga dataset.}

\item{version_number}{the version number of the taiga dataset}

\item{file_name}{the name of the taiga file}
}
\value{
a sparse matrix of type S4
}
\description{
download raw sparse matrix data files from taiga to R
}
