% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{extract_hugo_symbol_colnames}
\alias{extract_hugo_symbol_colnames}
\title{Extract hugo symbol column names from data matrices using "HGNC (ENTREZ)" format}
\usage{
extract_hugo_symbol_colnames(dat)
}
\arguments{
\item{dat:}{matrix or list of matrices}
}
\description{
Extract hugo symbol column names from data matrices using "HGNC (ENTREZ)" format
}
