% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDoseCurve.R
\name{fitDoseCurve}
\alias{fitDoseCurve}
\title{fitDoseCurve}
\usage{
fitDoseCurve(curve.data, upper_asym = 100, lower_asym = NA,
  robust = "median")
}
\arguments{
\item{curve.data}{data frame with viability ~ pert_effective_dose}

\item{upper_asym}{default set to 100}

\item{lower_asym}{default set to be floating}

\item{robust}{a character string specifying the rho function for robust estimation. 
Default is non-robust least squares estimation ("mean"). Available robust methods are: 
median estimation ("median"), least median of squares ("lms"), least trimmed squares 
("lts"), metric trimming ("trimmed"), metric winsorizing ("winsor") and Tukey's biweight ("tukey").}
}
\value{
A data.frame with fields for the IC50 ('ic50'), potency ('slope'),
}
\description{
Fit dose-response curves to each PRISM plate
Assumes a 4-parameter fit with the following parameterization:
f(x) = c + \\frac{d-c}{1+\\exp(b(\\log(x)-\\log(e)))}
where (e) is the inflection point of the curve, (b) is the slope, (c) is the lower asymptote, 
and (d) is the upper asymptote
}
