% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{get_cormat}
\alias{get_cormat}
\title{Compute correlation matrix for two matrices}
\usage{
get_cormat(mx1, mx2, Z.score = NA)
}
\arguments{
\item{mx1}{first input matrix, cannot be NA}

\item{mx2}{second input matrix, can be NA. When mx2 is NA, it will compute self correlation for mx1}

\item{Z.score}{one of strings "by mx1", by mx2", or "global", it will transform each correlation value to Z scores}
}
\description{
Compute correlation matrix for two matrices
}
