% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{get_effective_dose}
\alias{get_effective_dose}
\title{convert pert_dose to standard dose dilution
 @export to_df}
\usage{
get_effective_dose(pert_dose, effective_dose)
}
\description{
convert pert_dose to standard dose dilution
 @export to_df
}
