% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{get_gene_stats}
\alias{get_gene_stats}
\title{Helper function for computing gene level stats}
\usage{
get_gene_stats(X, y, stat_type = "t_stat", stat_trans = "none")
}
\arguments{
\item{X}{Samples by genes data matrix}

\item{y}{Phenotype vector}

\item{stat_type}{Type of stat c(t_stat, cor, mod_t_stat, pval, mod_pval, regr_coef)]}

\item{stat_trans}{Type of transformation c(square, abs, rank)}
}
\value{

}
\description{
Helper function for computing gene level stats
}
