% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{gsc_to_sparseMat}
\alias{gsc_to_sparseMat}
\title{Convert gene-set collection object to sparse matrix}
\usage{
gsc_to_sparseMat(gsc, gene_list)
}
\arguments{
\item{gsc}{gene set collection object}

\item{gene_list}{vector of gene names, specifying the set of genes included in analysis}
}
\value{
: sparse binary matrix of gene set memberships
}
\description{
Convert gene-set collection object to sparse matrix
}
