% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normLRT_test.R
\name{is.error}
\alias{is.error}
\title{Check error}
\usage{
is.error(x)
}
\arguments{
\item{x}{return from try() function}
}
\value{
boolean indicating if the call resulted in error
}
\description{
Check error
}
