% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_with_cor.R
\name{lm_with_cor}
\alias{lm_with_cor}
\title{ggduo module of scatter plot with linear fit}
\usage{
lm_with_cor(data, mapping, ..., method = "pearson")
}
\description{
ggduo module of scatter plot with linear fit
}
