% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normLRT_test.R
\name{LRT_init}
\alias{LRT_init}
\title{Function for fitting skew-t distribution using fixed nu param}
\usage{
LRT_init(vec, starting_nu)
}
\arguments{
\item{vec}{vector of values, at least 10 data points required (much more probably better)}

\item{starting_nu}{Some initial value of nu >= 2}
}
\value{
log likelihood of skewed t-dist fit
}
\description{
Function for fitting skew-t distribution using fixed nu param
}
