% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_network.R
\name{make_a_network}
\alias{make_a_network}
\title{Create network visualization}
\usage{
make_a_network(node_df, edge_df, directed = T, layout = layout_nicely,
  plot = T, detect_communities = T, detect_cliques = F)
}
\arguments{
\item{node_df}{data.frame of nodes in network}

\item{edge_df}{data.frame of edges in network}
}
\value{
An igraph object (see package \code{igraph} for more details)
}
\description{
Create network visualization
}
\details{
`node_df` must have a column called `id` that contains the node identifiers. `edge_df` must have at least
two columns labeled `from` and `to` respectively. Optional columns in `node_df` are `label`, `size`,
`color`, and `shape`
}
\examples{
nodes <- data.frame(id=LETTERS)
edges <- data.frame(from=LETTERS[sample(1:26, 10)], to=LETTERS[sample(1:26, 30, replace=T)])
make_a_network(nodes, edges)
}
