% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dep_report.R
\name{make_dep_report}
\alias{make_dep_report}
\title{Create html report assessing genetic and chemical dependencies related to specified cell line features}
\usage{
make_dep_report(CL_var, covars = NULL, test_dir = "left",
  report_title = "dep_report", used_gscs = c("GO_biological_process",
  "canonical", "CORUM"), out_dir = NULL, save_pref_deps = FALSE,
  nproc = 0)
}
\arguments{
\item{CL_var}{vector of cell line features. Can be either binary/logical specifying two groups, or continuous values. Must be named with CCLE_IDs}

\item{covars}{Optional named vector of covariates to be 'controlled for' in analysis}

\item{test_dir}{Direction of association testing. Default is left-tailed tests, which corresponds to stronger (more negative) dependencies when e.g. the 'in-group' is of interest in a two-group comparison. Options are ['left', 'right', and 'both']}

\item{report_title}{Name of report}

\item{used_gscs}{Name of gene set collections to use. Defaults to include GO_biological_process, canonical (from MSIGDB) and CORUM complex co-membership}

\item{out_dir}{Directory to generate report}

\item{save_pref_deps}{Should the table of preferential dep stats be saved to an rds file (default FALSE)}

\item{nproc}{Number of processes to use (0 for no multithreading). Default 0. Only affects gsa calcs with covariates at this stage}
}
\description{
Create html report assessing genetic and chemical dependencies related to specified cell line features
}
