% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_group_boxplot}
\alias{make_group_boxplot}
\title{Make boxplot}
\usage{
make_group_boxplot(df, group, yvar, min_per_group = 2,
  group_order = "increasing", show_points = TRUE)
}
\arguments{
\item{df:}{data frame}

\item{group:}{string name of variable to group by}

\item{yvar:}{string name of y variable}

\item{min_per_group:}{min samples per group [2]}

\item{group_order:}{how order groups. ('increasing' [default], 'decreasing', 'none')}

\item{show_points:}{whether or not to show individual data points [TRUE]}
}
\description{
Make boxplot
}
