% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{make_GSA_network}
\alias{make_GSA_network}
\title{Make network visualization of gene set stats}
\usage{
make_GSA_network(gene_set_stats, gsc, all_genes, top_N = 100,
  one_per_cluster = TRUE, min_cluster_size = 2, vertex.label.cex = 0.5)
}
\arguments{
\item{gsc}{Original gene set collection object}

\item{all_genes}{List of all genes used in original analysis}

\item{top_N}{number of top gene sets to visualize in the graph (ranked by stat)}

\item{one_per_cluster}{Whether to label just the top gene set per cluster (default TRUE)}

\item{min_cluster_size}{Minimum cluster size to include in the final plot (default 2)}

\item{vertex.label.cex}{Size of vertex labels (default 0.5)}

\item{gene_stats}{dataframe containing at minimum columns 'gene_set' and 'stat'}
}
\value{
igraph plot
}
\description{
Make network visualization of gene set stats
}
