% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Scatterplot comparing two variables with gene labels based on input subset of genes}
\usage{
make_scatter_plot(df, xvar, yvar, xlab = NULL, ylab = NULL,
  gene_field = "Gene", top_genes = 25, lab_size = 2,
  color_annot = NULL, show_stats = TRUE, make_plotly = FALSE)
}
\arguments{
\item{make_plotly}{}
}
\description{
Scatterplot comparing two variables with gene labels based on input subset of genes
}
