% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_dataset.R
\name{make_tidy_dataset}
\alias{make_tidy_dataset}
\title{Make a tidy dataset}
\usage{
make_tidy_dataset(mat_list, na_rows = TRUE, na_cols = FALSE,
  use_dims = NULL, dim_names = c("Gene", "Sample"))
}
\arguments{
\item{mat_list}{list of matrices (columns are genes, rows are samples)}

\item{na_rows}{if TRUE, add NAs for missing rows, otherwise they are removed}

\item{na_cols}{if TRUE, add NAs for missing columns, otherwise they are removed}

\item{use_dims}{either an integer or character vector of length 1}

\item{dim_names}{character vector of length two -
these become the column labels in the tidy dataset
for the rownames and colnames of the original matrices}
}
\value{
a tidy dataframe
}
\description{
Takes a list of matrices with similar row and column names
and runs tidyr::gather across each to make a "long" dataset
with a column of values for each matrix in the list.
}
\examples{
CCLE_GE <-  taigr::load.from.taiga(
                data.name="ccle-rnaseq-expression-genes",
                data.version = 3) \%>\%
             set_colnames(stringr::str_match(colnames(.), '(.+) ')[,2]) \%>\%
             t()
             
MUT_MIS <- taigr::load.from.taiga(
                data.name = 'ccle-mis-mut-binary-matrix',
                data.version = 1,
                transpose = T)
CN <- taigr::load.from.taiga(
                data.name="ccle-copy-number-variants-hgnc-mapped",
                data.version = 1)             
             
gene_feature_df <- cdsr::make_tidy_dataset(list(exp=CCLE_GE,
                                                mut=MUT_MIS,
                                                cn=CN),
                                           dim_names=c("Gene", "CellLine"))
head(gene_feature_df)
             

}
