% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{map_arxspan_to_ccle}
\alias{map_arxspan_to_ccle}
\title{Apply arxspan to ccle conversion to matrix or list of matrices (assuming cell line names are in the rownames)}
\usage{
map_arxspan_to_ccle(dat)
}
\arguments{
\item{dat:}{matrix or list of matrices}
}
\value{
matrix or list of matrices with rownames converted to CCLE ID
}
\description{
Apply arxspan to ccle conversion to matrix or list of matrices (assuming cell line names are in the rownames)
}
