% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_matrix.R
\name{melt_matrix}
\alias{melt_matrix}
\title{Convert a matrix into a data.frame}
\usage{
melt_matrix(mat, row_name, col_name, value_name)
}
\arguments{
\item{mat}{A matrix.}

\item{row_name}{String label for matrix row names}

\item{col_name}{String label for matrix column names}

\item{value_name}{String label for matrix entries}
}
\value{
A data.frame containing columns for \code{row_name}, \code{col_name}, and \code{value_name}
}
\description{
Convert a matrix into a data.frame
}
\examples{
MUT.DAM <- load.from.taiga(data.name = 'ccle-mut-data-binary-matrix', data.version = 1)
MUT.DAM.DF <- melt_matrix(MUT.DAM, row_name = "ccle_name", col_name = "gene", value_name="mutation")
}
