% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normLRT_test.R
\name{normLRT_test}
\alias{normLRT_test}
\title{Measure normality likelihood ratio statistic, as described in MacDonald et al Cell 2017}
\usage{
normLRT_test(vec)
}
\arguments{
\item{vec}{vector of values, at least 10 data points required (much more probably better)}
}
\value{
twice log likelihood ratio of skewed t-dist fit compared to guassian fit
}
\description{
Measure normality likelihood ratio statistic, as described in MacDonald et al Cell 2017
}
