% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_output_parser.R
\name{parse_atlantis_table}
\alias{parse_atlantis_table}
\title{parsing an ATLANTIS summary table}
\usage{
parse_atlantis_table(df, match.confounder = T, cutoff = 0)
}
\arguments{
\item{df}{The ATLANTIS summary table download from Tundra}

\item{match.confounder}{logic of whether to match the confounder-only model for each targetID}

\item{cutoff}{a numeric value between -1 and 1. When match.confounder is TRUE, the corrected Rsquared will be calculated and value below or equal to this cutoff will be filtered out}
}
\value{
if match.confounder is TRUE, A data frame in which each row contains the ATLANTIS model metrics of Rsquared and topFeatureas well as the model with confounder only as feature. If match.confounder is FALSE, no confounder models will be matched and the output also include a column of report url
}
\description{
parsing an ATLANTIS summary table
}
\details{
For generating the main table in the project dashboard, use match.confounder = T, cutoff = 0. For generating the table in the Biomarker tab in the compound dashboard, use match.confounder = F
}
\author{
Li Wang
}
