% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseCurve.R
\name{plot_sigmoid}
\alias{plot_sigmoid}
\title{plot_sigmoid}
\usage{
plot_sigmoid(data_dose_cpd, data_curved_cpd)
}
\arguments{
\item{data_dose_cpd}{data frame with viability ~ pert_effective_dose}

\item{data_curved_cpd}{output of fitDoseCurve, a data frame}
}
\description{
plot sigmoidal dose response curves
}
