% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{process_gctx_prism}
\alias{process_gctx_prism}
\title{run combat for batch correction}
\usage{
process_gctx_prism(file.name, exlucde.wells)
}
\arguments{
\item{file.name}{a character string of the path of the input file}

\item{exlucde.wells}{well to exclude, a vector of platewell_id, i.e. det_plate:pert_well}
}
\description{
run combat for batch correction
}
