% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cols_all_na.R
\name{remove_cols_all_na}
\alias{remove_cols_all_na}
\title{Remove columns from a matrix that contain all NA entries}
\usage{
remove_cols_all_na(x, indices = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{indices}{logical, return indices to remove instead}

\item{verbose}{logical, display number of rows removed}
}
\value{
the matrix with columns removed, unless indices is TRUE
}
\description{
Remove columns from a matrix that contain all NA entries
}
\examples{
mat <- matrix(c(NA, NA, 3, 4), nrow=2, ncol=2)
remove_cols_all_na(mat)
}
