% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rows_all_na.R
\name{remove_rows_all_na}
\alias{remove_rows_all_na}
\title{Remove rows from a matrix that are entirely NA}
\usage{
remove_rows_all_na(x, indices = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{indices}{logical, return indices to remove instead}

\item{verbose}{logical, display number of rows removed}
}
\value{
the matrix with rows removed, unless indices is TRUE in which case the
indices of the rows with all NA entries are returned
}
\description{
Remove rows from a matrix that are entirely NA
}
\examples{
mat <- matrix(c(NA, 2, NA, 4), nrow=2, ncol=2)
remove_rows_all_na(mat)
}
