% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rows_any_na.R
\name{remove_rows_any_na}
\alias{remove_rows_any_na}
\title{Remove rows from a matrix that have any NA entries}
\usage{
remove_rows_any_na(x, indices = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{indices}{logical, return indices to remove instead}

\item{verbose}{logical, display number of rows removed}
}
\value{
the matrix with rows removed, unless indices is TRUE
}
\description{
Remove rows from a matrix that have any NA entries
}
\examples{
mat <- matrix(c(NA, 2, 3, 4), nrow=2, ncol=2)
remove_rows_any_na(mat)
}
