% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{rev_factor}
\alias{rev_factor}
\title{run combat for batch correction}
\usage{
rev_factor(x)
}
\arguments{
\item{x}{a vector, can be numeric, character or factor}
}
\value{
a vector of factor with the default order of factor reversed
}
\description{
run combat for batch correction
}
