% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{run_combat}
\alias{run_combat}
\title{run combat for batch correction}
\usage{
run_combat(df_data, value.var, output.name = "norm_viab_combat")
}
\arguments{
\item{df_data}{data frame containing columns of "pert_type", "pert_mfc_id", "pert_effective_dose", "det_plate", "cell_line_id",  and "pool_id"}

\item{value.var}{colnames of the viability column}
}
\description{
run combat for batch correction
}
