% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{run_fGSEA}
\alias{run_fGSEA}
\title{Run GSEA using fgsea package, either with gene or label permutation null}
\usage{
<<<<<<< HEAD
run_fGSEA(X, y, gsc, perm_type = "label", nperm = 10000, min_set_size = 3,
  max_set_size = 300, stat_type = "t_stat", stat_trans = "none",
  gseaParam = 1, nproc = 0)
}
\arguments{
\item{X}{matrix of data with rows as cell lines and columns as genes (same gene names as in gsc object)}

\item{y}{vector of phenotypes}

\item{gsc}{gene-set collection object from GSEAbase}
=======
run_fGSEA(gsc, X = NULL, y = NULL, perm_type = "label", nperm = 10000,
  min_set_size = 3, max_set_size = 300, stat_type = "t_stat",
  stat_trans = "none", gseaParam = 1, nproc = 0, gene_stat = NULL)
}
\arguments{
\item{gsc}{gene-set collection object from GSEAbase}

\item{X}{matrix of data with rows as cell lines and columns as genes (same gene names as in gsc object)}

\item{y}{vector of phenotypes.}
>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a

\item{perm_type}{type of permutations to generate null ['label', or 'gene']}

\item{nperm}{number of permutations}

\item{min_set_size}{minimum gene set size}

\item{max_set_size}{max gene set size}

\item{stat_type}{if using gene-permutation, stat to use. Options ['t_stat', 'mod_t_stat', 'cor', 'regr_coef', 'pval', 'mod_pval']}

\item{stat_trans}{if using gene-permutation, stat transformation to apply Options ['none', 'square', 'abs', 'rank']}

\item{gseaParam:}{GSEA parameter value (power)}

\item{nproc:}{number of processes to run in parallel}
<<<<<<< HEAD
=======

\item{gene_stat:}{vector of pre-computed gene-level stats. Only works for 'gene-permutation' testing. IF provided X and y will be ignored}
>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a
}
\description{
Run GSEA using fgsea package, either with gene or label permutation null
}
