% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{run_GSAhyper}
\alias{run_GSAhyper}
\title{Run set overlap test using piano 'runGSAhyper' function}
\usage{
run_GSAhyper(hit_genes, all_genes, gsc, gsSizeLim = c(1, Inf),
  adjMethod = "BH")
}
\arguments{
\item{hit_genes}{list of genes scoring as hits}

\item{all_genes}{list of all genes tested}

\item{gsc}{gene set collection as gene-set collection object}

\item{gsSizeLim}{2x1 vector of size limits on the gene sets (defaults to c(1,Inf))}

\item{adjMethod}{Method for p-valuea adjustment (from available in p.adjust, default = 'BH')}
}
\value{
Table of stats for all gene sets in collection
}
\description{
Run set overlap test using piano 'runGSAhyper' function
}
