% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lm_stats_limma.R
\name{run_lm_stats_limma_group}
\alias{run_lm_stats_limma_group}
\title{Estimate linear-model stats for a matrix of data wrt a group of phenotype variables using limma with empirical Bayes moderated F-stats for p-values.
Note: Design matrix only models levels with at least 3 samples (the rest of samples are just modeled with the intercept term). Otherwise, limma hits errors}
\usage{
run_lm_stats_limma_group(mat, phenos, covars = NULL, weights = NULL,
  target_type = "Gene", limma_trend = FALSE)
}
\arguments{
\item{mat:}{Nxp data matrix with N cell lines and p genes}

\item{phenos:}{Nxq dataframe/matrix of independent variables.}

\item{covars:}{Optional Nxk matrix of covariates}

\item{weights:}{Optional N vector of precision weights for each data point}

\item{target_type:}{Name of the column variable in the data (default 'Gene')}

\item{limma_trend:}{Whether to fit an intensity trend with the empirical Bayes variance model}
}
\value{
: data frame of stats
}
\description{
Estimate linear-model stats for a matrix of data wrt a group of phenotype variables using limma with empirical Bayes moderated F-stats for p-values.
Note: Design matrix only models levels with at least 3 samples (the rest of samples are just modeled with the intercept term). Otherwise, limma hits errors
}
\examples{
gene.effect <- load.from.taiga(data.name='avana-broad-18q1-ec8b', data.version=2, data.file='gene_effect')
sample.info <- load.from.taiga(data.name='avana-broad-18q1-ec8b', data.version=2, data.file='sample_info')

covars <- sample.info \%>\%
  dplyr::select(CCLE_ID = cell_line, cas9_activity, cell_line_SSMD) \%>\%
  mutate(cas9_activity = as.numeric(cas9_activity)) \%>\%
  filter(!is.na(cas9_activity)) \%>\%
  column_to_rownames(var = 'CCLE_ID')
uvars <- sample.info \%>\%
  dplyr::select(CCLE_ID = cell_line, primary_tissue) \%>\%
  column_to_rownames(var = 'CCLE_ID')
res <- cdsr::run_lm_stats_limma_group(gene.effect, uvars, covars, weights = NULL, target_type = 'Gene')

}
