% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{run_npGSEA}
\alias{run_npGSEA}
\title{Run npGSEA analysis, testing for gene set enrichment of association with a vector across a collection of gene sets}
\usage{
<<<<<<< HEAD
run_npGSEA(X, y, gsc, covars = NULL, approx = "norm", min_set_size = 3,
=======
run_npGSEA(gsc, X, y, covars = NULL, approx = "norm", min_set_size = 3,
>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a
  max_set_size = 300, normalize_X = FALSE, n_top_genes = 0,
  top_dir = "left", nproc = 0)
}
\arguments{
<<<<<<< HEAD
=======
\item{gsc:}{Gene set collection object}

>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a
\item{X:}{Matrix of gene-level data. Rows are cell line, columns are genes}

\item{y:}{Vector of data to compare with dep_mat. Can be categorical or continuous}

<<<<<<< HEAD
\item{gsc:}{Gene set collection object}

=======
>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a
\item{covars:}{optional matrix of covariates to control for. This matrix is linearly projected out of both X and y before further analysis}

\item{approx:}{type of approximation to use [default 'norm', other options 'beta' and 'chiSq'. See npGSEA docs]}

\item{min_set_size:}{minimum size of gene sets to consider [default 3]}

\item{max_set_size:}{Max size of gene sets to consider [default 300]}

\item{normalize_X:}{whether to normalize scale of data in each column of the X matrix to have unit variance [default FALSE]}

\item{n_top_genes:}{Optional number of top contributing genes from each set to return [default 0]}

\item{top_dir:}{If using n_top_genes, which 'direction' of test statistic should genes be ranked by [default 'left']}

\item{nproc:}{How many processes to use (default 0, no multithreading). Only helps when using covariates}
}
\value{
: data frame of results. Below assumes you're using the normal approximation. See docs if you're using a different model
\itemize{
\item gene_set: name of gene set
\item enrich_stat: sum of gene-level test statistics over genes in set
\item norm_enrich_stat: mean of gene-level test stats over genes in set
\item set_size: number of genes in set
\item p_left/p_right/p_twoSided: P-values for testing whether gene stats are decreased/increased/either for genes in set
\item q_left/q_right/q_twoSided: FDR_adjusted p-values
}
}
\description{
Run npGSEA analysis, testing for gene set enrichment of association with a vector across a collection of gene sets
}
