% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_functions.R
\name{run_piano}
\alias{run_piano}
\title{Run gene set analysis using piano package}
\usage{
<<<<<<< HEAD
run_piano(X, y, gsc, method, stat_type, stat_trans = "none", nperm = 10000,
  nproc = 1, verbose = TRUE)
}
\arguments{
=======
run_piano(gsc, X = NULL, y = NULL, gene_stat = NULL, stat_dir = NULL,
  method, stat_type, stat_trans = "none", nperm = 10000, nproc = 1,
  verbose = TRUE)
}
\arguments{
\item{gsc}{gene set collection as gene-set collection object}

>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a
\item{X}{matrix of data with rows as cell lines and columns as genes}

\item{y}{phenotype vector}

<<<<<<< HEAD
\item{gsc}{gene set collection as gene-set collection object}
=======
\item{gene_stat}{Optional: pre-computed gene-level stats. If given, X and y will be ignored}

\item{stat_dir}{Optional: pre-computed gene-stat directions}
>>>>>>> ddf2fc1032aef791ed610209f7b8bba2a354d64a

\item{method}{GSA method to use. Can be ['fisher', 'stouffer', 'reporter', 'tailStrength', 'mean', 'median', 'page']}

\item{stat_type}{Type of statistic to use for testing enrichment. Can be ['t_stat', 'mod_t_stat', 'cor', 'cor_rank', 'pval', 'mod_pval']. For methods 'fisher', 'stouffer', 'reporter', and 'tailStrength', must be p}

\item{stat_trans}{Type of transformation applied to gene stat Can be ['none', 'square', 'rank', 'abs'].}

\item{nperm}{Number of permutations to run (gene-permutation)}

\item{nproc}{Number of threads to use (requires snowfall library)}

\item{verbose}{}
}
\description{
Run gene set analysis using piano package
}
