% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{set_diag_na}
\alias{set_diag_na}
\title{make diagonal value NA for self correlation matrix}
\usage{
set_diag_na(mx)
}
\arguments{
\item{cormat}{square matrix with diagnal values equal to one}
}
\description{
make diagonal value NA for self correlation matrix
}
