% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Conduct t-tests on data within a data.frame}
\usage{
t_test(df, group_vars, condition_var, value_var)
}
\arguments{
\item{df}{A data.frame}

\item{group_vars}{Vector of string names for columns by which to group}

\item{condition_var}{String name of column indicating population conditions}

\item{value_var}{String name of column containing value to be tested}
}
\value{
A tibble containing a plot for each distinct combination of group_vars
}
\description{
Conduct t-tests on data within a data.frame
}
\examples{
plot_tbl <- t_test(mtcars, group_vars="vs", condition_var="am", value_var="mpg")
plot_tbl \%>\% dplyr::filter(vs == 1) \%>\% magrittr::extract2("plot")
}
