% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{theme_custom}
\alias{theme_custom}
\title{a ggplot theme that removes grid, background color, and top and right borders}
\usage{
theme_custom()
}
\description{
a ggplot theme that removes grid, background color, and top and right borders
}
