% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Set publication theme on ggplot object}
\usage{
theme_Publication(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_family}{}
}
\description{
Set publication theme on ggplot object
}
\examples{
https://rpubs.com/Koundy/71792
}
