% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{to_df}
\alias{to_df}
\title{take rownames as the first column and convert to data frame
 @export to_df}
\usage{
to_df(mx, colname1)
}
\description{
take rownames as the first column and convert to data frame
 @export to_df
}
