% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_helpers.R
\name{to_mat}
\alias{to_mat}
\title{take first column as rownames and convert to matrix
 @export to_mat}
\usage{
to_mat(df)
}
\description{
take first column as rownames and convert to matrix
 @export to_mat
}
