% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_test.R
\name{corr_test}
\alias{corr_test}
\title{Calculate correlations between matrices, with significance measures, accounting for missing data}
\usage{
corr_test(x, y, adjust = "none")
}
\arguments{
\item{x}{matrix of values}

\item{y}{second matrix of values (or a vector)}

\item{adjust}{p-value correction method (fed into p.adjust).}
}
\value{
: for matrix-matrix correlations returns a list of matrices. For matrix-vector correlations, returns a data frame
with values r: correlations, n: number observations, t: t-stats, p: p-values, se: standard errors, q: q-values
}
\description{
Calculate correlations between matrices, with significance measures, accounting for missing data
}
